Refactoring extensions.
