+++ NeoSwift/Extensions/ArrayExtension.swift
@@ -1,12 +1,13 @@
 //
-//  Array+Util.swift
+//  ArrayExtension.swift
 //  NeoSwift
 //
-//  Created by Luís Silva on 26/09/17.
-//  Copyright © 2017 drei. All rights reserved.
+//  Created by Ricardo Kobayashi on 10/10/18.
+//  Copyright © 2018 O3 Labs Inc. All rights reserved.
 //
 
 import Foundation
+import CommonCrypto
 
 extension Array where Element == UInt8 {
     public var hexString: String {
@@ -46,6 +47,35 @@ extension Array where Element == UInt8 {
             result.append(self[i] ^ other[i])
         }
         return result
+    }
+    
+    public var base58EncodedString: String {
+        guard !self.isEmpty else { return "" }
+        return Base58.base58FromBytes(self)
+    }
+    
+    public var base58CheckEncodedString: String {
+        var bytes = self
+        let checksum = [UInt8](bytes.sha256.sha256[0..<4])
+        
+        bytes.append(contentsOf: checksum)
+        
+        return Base58.base58FromBytes(bytes)
+    }
+    
+    public var sha256: [UInt8] {
+        let bytes = self
+        
+        let mutablePointer = UnsafeMutablePointer<UInt8>.allocate(capacity: Int(CC_SHA256_DIGEST_LENGTH))
+        
+        CC_SHA256(bytes, CC_LONG(bytes.count), mutablePointer)
+        
+        let mutableBufferPointer = UnsafeMutableBufferPointer<UInt8>.init(start: mutablePointer, count: Int(CC_SHA256_DIGEST_LENGTH))
+        let sha256Data = Data(buffer: mutableBufferPointer)
+        
+        mutablePointer.deallocate()
+        
+        return sha256Data.bytes
     }
 }
