+++ NeoSwift/Models/Marketplace/Base58.swift
@@ -113,53 +113,3 @@ struct Base58 {
         return result
     }
 }
-
-extension Array where Element == UInt8 {
-    public var base58EncodedString: String {
-        guard !self.isEmpty else { return "" }
-        return Base58.base58FromBytes(self)
-    }
-    
-    public var base58CheckEncodedString: String {
-        var bytes = self
-        let checksum = [UInt8](bytes.sha256.sha256[0..<4])
-        
-        bytes.append(contentsOf: checksum)
-        
-        return Base58.base58FromBytes(bytes)
-    }
-}
-
-extension String {
-    public var base58EncodedString: String {
-        return [UInt8](utf8).base58EncodedString
-    }
-    
-    public var base58DecodedData: Data? {
-        let bytes = Base58.bytesFromBase58(self)
-        return Data(bytes)
-    }
-    
-    public var base58CheckDecodedData: Data? {
-        guard let bytes = self.base58CheckDecodedBytes else { return nil }
-        return Data(bytes)
-    }
-    
-    public var base58CheckDecodedBytes: [UInt8]? {
-        var bytes = Base58.bytesFromBase58(self)
-        guard 4 <= bytes.count else { return nil }
-        
-        let checksum = [UInt8](bytes[bytes.count-4..<bytes.count])
-        bytes = [UInt8](bytes[0..<bytes.count-4])
-        
-        let calculatedChecksum = [UInt8](bytes.sha256.sha256[0...3])
-        if checksum != calculatedChecksum { return nil }
-        
-        return bytes
-    }
-    
-    public var littleEndianHexToUInt: UInt {
-        return UInt(self.dataWithHexString().bytes.reversed().fullHexString, radix: 16)!
-    }
-    
-}
